Seperated Type Spec related functions from methods
