+++ model/Model.go
@@ -7,8 +7,6 @@ import (
 	"fmt"
 	"github.com/metalim/jsonmap"
 	"github.com/rs/zerolog/log"
-	"reflect"
-	"slices"
 	"strings"
 )
 
@@ -23,92 +21,24 @@ func Create(name string, db *sql.DB, template string, JSON *jsonmap.Map) Model {
 	return Model{Name: name, db: db, template: template, json: JSON}
 }
 
-// Converts maps  into Array
-func MapToArray(s interface{}) ([]string, error) {
-	if s == nil {
-		return nil, errors.New("Nil value Passed")
-	}
-	switch v := s.(type) {
-	case *jsonmap.Map:
-		if v == nil {
-			return nil, errors.New("Nil value passed")
-		}
-		var values = make([]string, 0)
-
-		val := v.Values()
-		for i := 0; i < len(val); i++ {
-			str := fmt.Sprintf("%v", val[i])
-			values = append(values, str)
-		}
-		return values, nil
-	default:
-		return nil, errors.New("unsupported type")
-	}
-}
-func capitalize(s string) string {
-	if len(s) == 0 {
-		return s
-	}
-	return string(s[0]-'a'+'A') + s[1:]
-}
-
-func generateStructFromJsonMap(f jsonmap.Map) reflect.Type {
-	fields := make([]reflect.StructField, 0, len(f.Keys()))
-	for _, Name := range f.Keys() {
-		x, ok := f.Get(Name)
-		if !ok {
-			log.Fatal().Msg("Something went wrong ")
-		} else {
-			var T any
-			switch x {
-			case "string":
-				T = ""
-			case "integer":
-				T = 0
-			default:
-				log.Fatal().Msg("Unrecognized type in JSON template")
-			}
-			field := reflect.StructField{
-				Name: capitalize(Name),
-				Type: reflect.TypeOf(T),
-			}
-			fields = append(fields, field)
-
-		}
-	}
-	structType := reflect.StructOf(fields)
-	return structType
-}
-
-// fills out the query template with data from the json
+// Fills out the query template with data from the json
 func (m Model) Querybuilder(x []byte) (string, error) {
 	json1 := jsonmap.New()
-	var Keys = make([]string, 0)
-	var TKeys = make([]string, 0)
 
 	if len(x) == 0 {
 		log.Warn().Msg("Empty JSON template so Query is sent as is.")
 		return m.template, nil
 	}
+
 	err := json.Unmarshal(x, json1)
 	if err != nil {
 		return "", errors.New("failed to decode JSON data: " + err.Error())
 	}
 
-	err = json.Unmarshal(x, json1)
-	if err != nil {
-		return "", errors.New("failed to decode JSON data: " + err.Error())
-	}
-	val := json1.Keys()
-	for i := 0; i < len(val); i++ {
-		str := fmt.Sprintf("%v", val[i])
-		Keys = append(Keys, capitalize(str))
-	}
 	T := generateStructFromJsonMap(*m.json)
-	for i := 0; i < T.NumField(); i++ {
-		TKeys = append(TKeys, T.Field(i).Name) //
-	}
-	if slices.Equal(TKeys, Keys) {
+
+	//Check if request matches json specification
+	if matchesKeys(json1.Keys(), T) {
 		arrayData, err := MapToArray(json1)
 		if err != nil {
 			return "", err
@@ -128,14 +58,15 @@ func (m Model) Querybuilder(x []byte) (string, error) {
 		return fmt.Sprintf(m.template, args...), nil
 
 	} else {
-		log.Warn().Msg("JSON request does not match spec")
-		return "", errors.New("JSON request does not match spec")
+		err := "JSON request does not match spec"
+		log.Warn().Err(errors.New(err)).Msg("Malformed JSON")
+		return "", errors.New(err)
 	}
 
 }
 
+// Queries the database
 func (m Model) Query(query string) (*sql.Rows, error) {
-	fmt.Println(query)
 	rows, err := m.db.Query(query)
 	if err != nil {
 		return nil, err
